Refactor tests; assert on KeepOrdering failures.
